Lesson 7: Kotlin Singletons and Retrofit https://codingwithmitch.com/courses/model-view-intent-mvi-architecture/kotlin-singletons-and-retrofit/
Made our ApiService for Mitch's OpenApi server. Made our RetrofitBuilder object which holds references to a lazy Retrofit.Builder and a lazy ApiService. If we  were making a real app, we might use Dagger to make this a Singleton, and maybe scope it as well. But with things as simple as they are, a simple Kotlin object notation works just fine.